refactor: use reduct for dependency injection
