+++ package.json
@@ -46,6 +46,7 @@
     "moment": "^2.10.2",
     "oer-utils": "^1.3.4",
     "promise-retry": "^1.1.1",
+    "reduct": "^3.1.0",
     "riverpig": "^1.1.0",
     "sequelize": "^3.27.0",
     "through2": "^2.0.1",
