+++ src/services/routing-table.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const PrefixMap = require('../routing/prefix-map')
+
+class RoutingTable extends PrefixMap {}
+
+module.exports = RoutingTable
