--- test/data/tradingPairs.json
@@ -1,27 +1,0 @@
-[
-  [
-    "usd-ledger",
-    "eur-ledger"
-  ], [
-    "eur-ledger",
-    "usd-ledger"
-  ], [
-    "cad-ledger",
-    "eur-ledger"
-  ], [
-    "usd-ledger",
-    "cad-ledger"
-  ], [
-    "cad-ledger",
-    "usd-ledger"
-  ], [
-    "eur-ledger",
-    "cad-ledger"
-  ], [
-    "usd-ledger",
-    "cny-ledger"
-  ], [
-    "cny-ledger",
-    "usd-ledger"
-  ]
-]
